refactor(shared): dedupe frontmatter parsing and duration formatting
- skills/_engine/frontmatter.ts is the single parser; agents/_engine consumes it (was a byte-for-byte copy)
- format-duration shared by the journal and round-outcome renderers
