+++ src/business/sprint/render-journal-entry.ts
@@ -1,5 +1,6 @@
 import type { IsoTimestamp } from '@src/domain/value/iso-timestamp.ts';
 import type { LearningEntry } from '@src/domain/signal.ts';
+import { formatDuration } from '@src/business/_shared/format-duration.ts';
 import { neutralizeProseHeadings } from '@src/business/sprint/journal-sanitize.ts';
 import { renderSectionHeader } from '@src/business/sprint/journal-structure.ts';
 
@@ -151,28 +152,6 @@ export interface JournalEntryInput {
 }
 
 const EM_DASH = '—';
-
-const MS_PER_SECOND = 1000;
-const MS_PER_MINUTE = 60 * MS_PER_SECOND;
-const MS_PER_HOUR = 60 * MS_PER_MINUTE;
-
-/** Human-readable duration. Mirrors the `render-round-outcome` formatter. */
-const formatDuration = (ms: number | undefined): string => {
-  if (ms === undefined) return EM_DASH;
-  if (ms < 0) return `${String(ms)}ms`;
-  if (ms < MS_PER_SECOND) return `${String(ms)}ms`;
-  if (ms < MS_PER_MINUTE) {
-    return `${String(Math.floor(ms / MS_PER_SECOND))}s`;
-  }
-  if (ms < MS_PER_HOUR) {
-    const minutes = Math.floor(ms / MS_PER_MINUTE);
-    const seconds = Math.floor((ms % MS_PER_MINUTE) / MS_PER_SECOND);
-    return seconds > 0 ? `${String(minutes)}m ${String(seconds)}s` : `${String(minutes)}m`;
-  }
-  const hours = Math.floor(ms / MS_PER_HOUR);
-  const minutes = Math.floor((ms % MS_PER_HOUR) / MS_PER_MINUTE);
-  return minutes > 0 ? `${String(hours)}h ${String(minutes)}m` : `${String(hours)}h`;
-};
 
 const SHA_DISPLAY_LENGTH = 7;
