+++ src/integration/ai/skills/_engine/parse-skill.ts
@@ -2,10 +2,11 @@
  * Shared SKILL.md parsing helpers — the canonical frontmatter split + naive-YAML reader + body
  * extraction that every {@link SkillSource} backed by on-disk `SKILL.md` folders consumes.
  *
- * Extracted from `bundled/source.ts` so the bundled source and the operator source
- * (`operator/source.ts`) share one parse implementation byte-for-byte. The frontmatter we
- * author is always flat `key: value` lines, so a full YAML parser is overkill (and adds a dep);
- * a real YAML lib lands only when a skill needs nested / multiline frontmatter.
+ * The generic split/YAML/error-code primitives live in `./frontmatter.ts` — shared byte-for-byte
+ * with `agents/_engine/parse-agent-definition.ts`, which reaches into this module's sibling via
+ * the documented cross-concept `_engine`-to-`_engine` import seam (both formats use the same
+ * flat `key: value` frontmatter shape). This module re-exports `errorCode` so existing skill-
+ * source callers (`bundled/source.ts`, `operator/source.ts`, `phase/*.ts`) are unaffected.
  *
  * `parseSkill` validates against {@link SkillFrontmatterSchema} and asserts frontmatter `name`
  * matches the on-disk folder name per the Agent Skills spec. The `label` parameter tailors the
@@ -14,60 +15,11 @@
 
 import { Result } from '@src/domain/result.ts';
 import { StorageError } from '@src/domain/value/error/storage-error.ts';
+import { errorCode, parseSimpleYaml, splitFrontmatter } from '@src/integration/ai/skills/_engine/frontmatter.ts';
 import type { Skill } from '@src/integration/ai/skills/_engine/skill.ts';
 import { SkillFrontmatterSchema } from '@src/integration/ai/skills/_engine/skill.ts';
 
-/**
- * Split a SKILL.md body into frontmatter + content. The frontmatter block is the first
- * `---` … `---` pair starting at the file's first non-whitespace line; everything after the
- * closing fence is the body. Returns the body verbatim when no frontmatter is present —
- * callers then validate frontmatter separately.
- *
- * @public
- */
-export const splitFrontmatter = (raw: string): { readonly frontmatter: string; readonly body: string } => {
-  const trimmed = raw.replace(/^\uFEFF/u, ''); // strip UTF-8 BOM
-  if (!trimmed.startsWith('---')) return { frontmatter: '', body: trimmed };
-  const closing = trimmed.indexOf('\n---', 3);
-  if (closing === -1) return { frontmatter: '', body: trimmed };
-  const frontmatter = trimmed.slice(3, closing).trim();
-  const afterClose = trimmed.slice(closing + 4); // skip "\n---"
-  // Strip every blank line after the closing fence so the body is clean. Stripping only ONE
-  // line-end would keep the standard blank separator line inside the body, and each
-  // parse → render round-trip (render re-inserts the separator) would grow the file by one
-  // blank line.
-  const body = afterClose.replace(/^(?:\r?\n)+/, '');
-  return { frontmatter, body };
-};
-
-/**
- * Naive YAML key:value parser — keys are simple identifiers, values are strings or single-quoted
- * strings without escapes. Frontmatter we control is always this shape, so a full YAML parser
- * is overkill (and adds a dep). Multiline / nested YAML is rejected via schema validation.
- *
- * @public
- */
-export const parseSimpleYaml = (input: string): Record<string, string> => {
-  const result: Record<string, string> = {};
-  for (const line of input.split('\n')) {
-    const stripped = line.trim();
-    if (stripped.length === 0 || stripped.startsWith('#')) continue;
-    const colon = stripped.indexOf(':');
-    if (colon === -1) continue;
-    const key = stripped.slice(0, colon).trim();
-    let value = stripped.slice(colon + 1).trim();
-    if (value.startsWith('"') && value.endsWith('"')) value = value.slice(1, -1);
-    else if (value.startsWith("'") && value.endsWith("'")) value = value.slice(1, -1);
-    result[key] = value;
-  }
-  return result;
-};
-
-/** Narrow an unknown caught value to a Node `fs` error code without leaking `any`. @public */
-export const errorCode = (cause: unknown): string | undefined =>
-  typeof cause === 'object' && cause !== null && 'code' in cause && typeof cause.code === 'string'
-    ? cause.code
-    : undefined;
+export { errorCode };
 
 /**
  * Parse an already-read SKILL.md body into the canonical {@link Skill} record. Split from the
